Refine black frame threshold normalization
